Refactor countries filtering
Brings the filtering to the top level, outside the useEffect and always
applies whatever filtering criteria are active (wishlist, vis,ited,
and/or search) to the list of all countries.

This way we remove the need for 2 of the states on the component -
filtered list and the no results boolean, while keeping the allCountries
list intact.

This could be optimized in terms of performance by turning those 3
filter operations into one, but I think it's more readable this way and
should be fine anyway!